api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.1
redis==5.0.7
dnspython==2.7.0
python-dotenv==1.0.1
structlog==24.1.0
orjson==3.10.7
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.23.8
pydantic-settings==2.6.1
email-validator==2.2.0
